DDP-5802 - refactor FormInstanceCreator
